requirements.txt
cryptography==41.0.4
fastapi==0.103.2
pydantic==2.4.2
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
uvicorn==0.23.2
